% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Outlier_after_log_transformation.R
\name{Outlier_after_log_transformation}
\alias{Outlier_after_log_transformation}
\title{Outliers summary after log transformation – an extension of Outlier_wo_data_transformation function}
\usage{
Outlier_after_log_transformation(data1, columnn1, first_function)
}
\arguments{
\item{data1}{columnn1, first_function}
}
\description{
Summary of dataframe of Oultiers nos in Lower and upper bound before and after log transformation
}
\examples{
 Outlier_after_log_transformation(Boston, columnn1 ((User can use this parameter as it is)) , first_function (call the function “Outlier_wo_data_transformation))

}
