% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Outlier_wo_data_transformation.R
\name{Outlier_wo_data_transformation}
\alias{Outlier_wo_data_transformation}
\title{Outliers without any data transformation}
\usage{
Outlier_wo_data_transformation(data, columnn)
}
\arguments{
\item{data}{columnn}
}
\description{
Summary of dataframe of Oultiers nos in Lower and upper bound without any data transformation
}
\examples{
 Outlier_wo_data_transformation(Boston, columnn (user can pass this parameter as it is))

}
