% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentage_missing_values.R
\name{percentage_missing_values}
\alias{percentage_missing_values}
\title{Missing values summary}
\usage{
percentage_missing_values(data)
}
\arguments{
\item{data}{}
}
\description{
Printing dataframe of total missing values and it's percentage
}
\examples{
 percentage_missing_values("Boston")

}
