% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Statplot_all_numerics.R
\name{Statplot_all_numerics}
\alias{Statplot_all_numerics}
\title{Statistical plots for all numeric variables}
\usage{
Statplot_all_numerics(data, columnn, exportfileto = " ")
}
\arguments{
\item{data}{columnn (user can give this parameter as it is) exportfile}
}
\description{
Plot box plot and histogram for all numeric variables
}
\examples{
 Statplot_all_numerics ("Boston", columnn, exportfileto = "new directory")

}
