% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Statplot_all_numerics_cum_categorical.R
\name{Statplot_all_numerics_cum_categorical}
\alias{Statplot_all_numerics_cum_categorical}
\title{Numeric and Categoric statistic plots}
\usage{
Statplot_all_numerics_cum_categorical(data, columnn, exportfileto = " ")
}
\arguments{
\item{data}{columnn (user can give this parameter as it is) exportfileto}
}
\description{
Plot box plot and histogram for numeric variables & Pie chart and Barplot for categorical variables
}
