% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Statplots_aggregate_images_1.R
\name{Statplots_aggregate_images_1}
\alias{Statplots_aggregate_images_1}
\title{png files aggregation}
\usage{
Statplots_aggregate_images_1(data, variable = c(), exportfileto = " ",
  newfilepath = "")
}
\arguments{
\item{data}{variable (user interest variable index) exportfile newfilepath}
}
\description{
Moving all the png files from one folder to another folder
}
\examples{
 Statplots_aggregate_images_1("Boston", variable = c(1,2,3), pathfrom, pathto)

}
