% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Statplots_selected_numeric_variable_1.R
\name{Statplots_selected_numeric_variable_1}
\alias{Statplots_selected_numeric_variable_1}
\title{Statistical plots for user interest variables}
\usage{
Statplots_selected_numeric_variable_1(data, variable = c(),
  exportfileto = " ")
}
\arguments{
\item{data}{variable (vector of user interest variable index) exportfileto (current directory)}
}
\description{
Plot box plot and histogram for user input numeric variables
}
\examples{
 Statplots_selected_numeric_variable_1 ("Boston", variable = c(1,2,3), exporfileto = "Directory name")

}
